Create digital timer
The timer should count up 10 seconds, increments every 10 ms. It
should have digits change to red when it gets to 10 seconds, and it
should not count past 10 seconds.

Implement a start and a reset button, and styled them. the start
button is disabled when the user clicks it, and not enabled until
the timer finishes.